Naming convention for Components and general cleanup
