+++ src/Climma/index.tsx
@@ -1,23 +1,18 @@
 import React from 'react'
-import { Box, Text, StdinContext, Color } from 'ink'
+import { Box, Color } from 'ink'
 import Spinner from 'ink-spinner'
 
-import { getWeatherByCity } from '../api'
-import getConfig from '../config'
-import temperatureColor from '../temperature'
-import getCondition, { Conditions } from '../conditions'
-import direction from '../direction'
-import Ascii from '../conditions/ascii'
+// Api 
+import { getWeatherByCity, WeatherData } from '../api'
 
-type WeatherData = {
-  weatherId?: number;
-  windDegree?: number;
-  windSpeed?: number;
-  minTemperature?: number;
-  maxTemperature?: number;
-  rain?: number;
-  city?: string;
-}
+// Entities
+import getConfig from '../config'
+import getCondition from '../conditions'
+
+// Components
+import Temperature from '../Temperature'
+import Direction from '../Direction'
+import Ascii from '../conditions/Ascii'
 
 interface State {
   weather: WeatherData;
@@ -26,15 +21,6 @@ interface State {
 
 interface Props {
   readonly city: string;
-}
-
-const Temperature = ({temp}: {temp?: number}) => {
-  if (temp === undefined) {
-    return <Color ansi256={203}>{temp} °C</Color>
-  }
-
-  const n = temperatureColor(temp)
-  return <Color ansi256={n}>{temp} °C</Color>
 }
 
 class Climma extends React.Component<Props> {
@@ -86,7 +72,7 @@ class Climma extends React.Component<Props> {
             <Box>
               <Temperature temp={weather.minTemperature}/> - <Temperature temp={weather.maxTemperature}/>
             </Box>
-            <Box>{direction(weather.windDegree)} {weather.windSpeed} km/h</Box>
+            <Box><Direction deg={weather.windDegree} /> {weather.windSpeed} km/h</Box>
             <Box marginBottom={1}>{weather.rain}mm</Box>
             <Box>
               <Color green>Weather generated by <Color bold>climma</Color></Color>
