+++ src/Temperature/index.tsx
@@ -1,4 +1,5 @@
-import chalk from 'chalk'
+import React from 'react'
+import { Color } from 'ink'
 
 import { getColorInRange } from '../util'
 
@@ -12,4 +13,13 @@ const temperatureColor = (temp: number) => {
   return n
 }
 
-export default temperatureColor
+const Temperature = ({temp}: {temp?: number}) => {
+  if (temp === undefined) {
+    return <Color ansi256={203}>{temp} °C</Color>
+  }
+
+  const n = temperatureColor(temp)
+  return <Color ansi256={n}>{temp} °C</Color>
+}
+
+export default Temperature
